requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
redis
python-multipart
